# OatMilk.Backend.Api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OatMilk.Backend.Api",
    "version": "v1"
  },
  "paths": {
    "/Character/full": {
      "get": {
        "tags": [
          "Character"
        ],
        "summary": "Get a paginated, filtered and sorted list of all existing characters, with all details.",
        "parameters": [
          {
            "name": "SearchByName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortColumnName",
            "in": "query",
            "description": "Name of the column to sort by.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortAscending",
            "in": "query",
            "description": "Whether or not to sort in ascending order.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "PageIndex",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponsePageResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponsePageResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponsePageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Character"
        ],
        "summary": "Create a character.",
        "requestBody": {
          "description": "Details of new character.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Character/full/{identifier}": {
      "get": {
        "tags": [
          "Character"
        ],
        "summary": "Get an existing character by its human-readable identifier.",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "description": "Human readable unique identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Character/full/{id}": {
      "put": {
        "tags": [
          "Character"
        ],
        "summary": "Update an existing character by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Details of new character.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Character"
        ],
        "summary": "Delete an existing character by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CharacterAbilityScoreProficiencyRequest": {
        "required": [
          "abilityScoreId",
          "id"
        ],
        "type": "object",
        "properties": {
          "abilityScoreId": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "id": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "maxLength": 64,
            "type": "string",
            "nullable": true
          },
          "proficient": {
            "type": "boolean"
          },
          "expertise": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CharacterAbilityScoreProficiencyResponse": {
        "required": [
          "expertise",
          "id",
          "name",
          "proficient"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "proficient": {
            "type": "boolean"
          },
          "expertise": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CharacterAbilityScoreRequest": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "maxLength": 64,
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "proficient": {
            "type": "boolean"
          },
          "expertise": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CharacterAbilityScoreResponse": {
        "required": [
          "expertise",
          "id",
          "name",
          "proficiencies",
          "proficient",
          "value"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "proficient": {
            "type": "boolean"
          },
          "expertise": {
            "type": "boolean"
          },
          "proficiencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterAbilityScoreProficiencyResponse"
            }
          }
        },
        "additionalProperties": false
      },
      "CharacterAttributeRequest": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "maxLength": 64,
            "type": "string",
            "nullable": true
          },
          "currentValue": {
            "type": "integer",
            "format": "int32"
          },
          "defaultValue": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CharacterAttributeResponse": {
        "required": [
          "currentValue",
          "defaultValue",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "currentValue": {
            "type": "integer",
            "format": "int32"
          },
          "defaultValue": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CharacterDescriptionRequest": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "maxLength": 64,
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CharacterDescriptionResponse": {
        "required": [
          "id",
          "name",
          "value"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CharacterLevelResponse": {
        "required": [
          "experience",
          "level",
          "nextLevelExperienceRequirement",
          "previousLevelExperienceRequirement"
        ],
        "type": "object",
        "properties": {
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "experience": {
            "type": "integer",
            "format": "int32"
          },
          "previousLevelExperienceRequirement": {
            "type": "integer",
            "format": "int32"
          },
          "nextLevelExperienceRequirement": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CharacterRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterAttributeRequest"
            },
            "description": "A list of all attributes. If this is null, character's attributes won't be updated.",
            "nullable": true
          },
          "abilityScores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterAbilityScoreRequest"
            },
            "description": "A list of all attributes. If this is null, character's ability scores won't be updated.",
            "nullable": true
          },
          "abilityScoreProficiencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterAbilityScoreProficiencyRequest"
            },
            "description": "A list of all ability score proficiencies. If this is null, character's proficiencies won't be updated.\r\nAll requests here will be parented under existing ability scores.",
            "nullable": true
          },
          "descriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterDescriptionRequest"
            },
            "description": "A list of all attributes. If this is null, character's descriptions won't be updated.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CharacterResponse": {
        "required": [
          "abilityScores",
          "attributes",
          "createdDateTimeUtc",
          "descriptions",
          "id",
          "identifier",
          "level",
          "name",
          "updatedDateTimeUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          },
          "createdDateTimeUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDateTimeUtc": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "$ref": "#/components/schemas/CharacterLevelResponse"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterAttributeResponse"
            }
          },
          "abilityScores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterAbilityScoreResponse"
            }
          },
          "descriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterDescriptionResponse"
            }
          }
        },
        "additionalProperties": false
      },
      "CharacterResponsePageResponse": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "stackTrace": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}